Tried to improve the font; added effects.
I like this font better, since it doesn't have the same scaling artifacts and I think feels sci-fi. The license only requires attribution, as far as I understand CC-BY-SA. The effects are something I tinkered with a bit; there's about one or two per screen so it isn't as overwhelming.